Share common tests between component and hook.
